+++ libcnb-data/src/lib.rs
@@ -13,6 +13,7 @@ pub mod build;
 pub mod build_plan;
 pub mod buildpack;
 pub mod buildpack_plan;
+pub mod exec_d;
 pub mod launch;
 pub mod layer;
 pub mod layer_content_metadata;